requirement.txt
Confluence-py==1.0.1
python-jenkins==0.4.16 
requests==2.6.1
requests-kerberos==0.11.0
httplib2==0.10.3
google-api-python-client==1.6.7
python-bugzilla==2.1.0
paramiko==2.1.2
scp==0.10.2
